test(lovasz): check the loss against its definition, not the old loop
Hand-computed two-class example, perfect and completely wrong predictions,
the [0, 1] bound, absent classes, ignored points, point-order invariance and a
finite-difference gradient check replace the comparison against the previous
implementation. _lovasz_grad no longer forces float32, so the loss also runs in
float64, which the gradient check needs.

+++ autoware_ml/losses/segmentation3d/lovasz.py
@@ -36,8 +36,10 @@ def _lovasz_grad(sorted_ground_truth: torch.Tensor) -> torch.Tensor:
     """
     point_count = len(sorted_ground_truth)
     foreground_total = sorted_ground_truth.sum()
-    intersection = foreground_total - sorted_ground_truth.float().cumsum(0)
-    union = foreground_total + (1 - sorted_ground_truth).float().cumsum(0)
+    # Keep the caller's floating dtype rather than forcing float32, so the loss
+    # also runs in float64 (gradcheck) and mixed precision without a dtype clash.
+    intersection = foreground_total - sorted_ground_truth.cumsum(0)
+    union = foreground_total + (1 - sorted_ground_truth).cumsum(0)
     jaccard = 1.0 - intersection / union
     if point_count > 1:
         jaccard[1:point_count] = jaccard[1:point_count] - jaccard[0:-1]
@@ -134,9 +136,7 @@ class LovaszLoss(_Loss):
             return (probabilities * 0.0).sum()
 
         losses = []
-        # One device->host copy for the class list; iterating the CUDA tensor
-        # directly would synchronise once per class, and every class listed here
-        # is present, so the old per-class emptiness check was redundant.
+        # `.tolist()` synchronized once. Omitting it would sync per iteration.
         for class_index in labels.unique().tolist():
             foreground = (labels == class_index).type_as(probabilities)
             class_errors = (foreground - probabilities[:, class_index]).abs()
